AmongUsBotWithC#/requirements.txt
pytesseract
selenium
Pillow
mss
opencv-python
discord
docutils